requirements.txt
boto3>=1.28.0
python-dotenv>=1.0.0
# Kafka producer (kafka/producer.py)
kafka-python>=2.0.2
requests>=2.28.0
